Refine image generation script

+++ scripts/prepare_raspbian_piEye_image.bash
@@ -20,23 +20,31 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+# Script location
+this_path="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+base_path="$( dirname "${this_path}" )"
+scripts_path="${base_path}/scripts/"
+echo "Base path at ${base_path}"
+echo "scripts path at ${scripts_path}"
+
 # Change these
 root_password_clear="correct horse battery staple"
-pi_password_clear="wrong cart charger paperclip"
+pi_password_clear="supersecret"
 image_to_download="https://downloads.raspberrypi.org/raspbian_lite_latest"
+github_repo="https://github.com/MemDynLab/BeholderPi.git"
 
 # Get SHA-256 for the lite image
 # Lite is currently the third item on the download page
 checksum="$(wget --quiet https://www.raspberrypi.org/downloads/raspbian/ -O - | egrep -m 3 'SHA-256' | awk -F '<|>' '{i++}i==3{print $9}')"
 
-sdcard_mount="/mnt/sdcard"
+sdcard_mount="/mnt/beholderpi_sdcard"
 public_key_file="id_ed25519.pub"
 
-if [ ! -e "${public_key_file}" ]
+if [[ ! -e "${public_key_file}" ]]
 then
     echo "Can't find the public key file \"${public_key_file}\""
     echo "You can create one using:"
-    echo "   ssh-keygen -t ed25519 -f ./id_ed25519 -C \"Raspberry Pi keys\""
+    echo "   ssh-keygen -t ed25519 -f ./id_ed25519 -C \"BeholderPi keys\""
     exit
 fi
 
@@ -45,11 +53,11 @@ wget --continue ${image_to_download} -O raspbian_lite_image.zip
 
 echo "Checking the SHA-256 of the downloaded image matches \"${checksum}\""
 
-if [ $( sha256sum raspbian_lite_image.zip | grep ${checksum} | wc -l ) -eq "1" ]
+if [[ $( sha256sum raspbian_lite_image.zip | grep ${checksum} | wc -l ) -eq "1" ]]
 then
-    echo "The checksums matche"
+    echo "The checksums match"
 else
-    echo "The checksums did not match"
+    echo "The checksums do not match"
     exit 1
 fi
 
@@ -60,28 +68,30 @@ mkdir ${sdcard_mount}
 extracted_image=$( 7z l raspbian_lite_image.zip | awk '/-raspbian-/ {print $NF}' )
 echo "The name of the image is \"${extracted_image}\""
 
-7z x raspbian_lite_image.zip
+# to overwrite, -aoa, to skip, -aos
+7z x raspbian_lite_image.zip -aos
 
-if [ ! -e ${extracted_image} ]
+if [[ ! -e ${extracted_image} ]]
 then
     echo "Can't find the image \"${extracted_image}\""
     exit
 fi
 
+echo ""
 echo "Mounting the sdcard boot disk"
 unit_size=$(fdisk --list --units  "${extracted_image}" | awk '/^Units/ {print $(NF-1)}')
 start_boot=$( fdisk --list --units  "${extracted_image}" | awk '/W95 FAT32/ {print $2}' )
 offset_boot=$((${start_boot} * ${unit_size})) 
 mount -o loop,offset="${offset_boot}" "${extracted_image}" "${sdcard_mount}"
-ls -al /mnt/sdcard
-if [ ! -e "${sdcard_mount}/kernel.img" ]
+ls -al ${sdcard_mount}
+if [[ ! -e "${sdcard_mount}/kernel.img" ]]
 then
     echo "Can't find the mounted card\"${sdcard_mount}/kernel.img\""
     exit
 fi
 
 touch "${sdcard_mount}/ssh"
-if [ ! -e "${sdcard_mount}/ssh" ]
+if [[ ! -e "${sdcard_mount}/ssh" ]]
 then
     echo "Can't find the ssh file \"${sdcard_mount}/ssh\""
     exit
@@ -89,21 +99,23 @@ fi
 
 umount "${sdcard_mount}"
 
-echo "Mounting the sdcard root disk"
+echo ""
+echo "Mounting the SD card root disk"
 unit_size=$(fdisk --list --units  "${extracted_image}" | awk '/^Units/ {print $(NF-1)}')
 start_boot=$( fdisk --list --units  "${extracted_image}" | awk '/Linux/ {print $2}' )
 offset_boot=$((${start_boot} * ${unit_size})) 
 mount -o loop,offset="${offset_boot}" "${extracted_image}" "${sdcard_mount}"
-ls -al /mnt/sdcard
+ls -al ${sdcard_mount}
 
-if [ ! -e "${sdcard_mount}/etc/shadow" ]
+if [[ ! -e "${sdcard_mount}/etc/shadow" ]]
 then
     echo "Can't find the mounted card\"${sdcard_mount}/etc/shadow\""
     exit
 fi
 
+# Security and Access
+echo ""
 echo "Change the passwords and sshd_config file"
-
 root_password="$( python3 -c "import crypt; print(crypt.crypt('${root_password_clear}', crypt.mksalt(crypt.METHOD_SHA512)))" )"
 pi_password="$( python3 -c "import crypt; print(crypt.crypt('${pi_password_clear}', crypt.mksalt(crypt.METHOD_SHA512)))" )"
 sed -e "s#^root:[^:]\+:#root:${root_password}:#" "${sdcard_mount}/etc/shadow" -e  "s#^pi:[^:]\+:#pi:${pi_password}:#" -i "${sdcard_mount}/etc/shadow"
@@ -115,13 +127,51 @@ cat ${public_key_file} >> "${sdcard_mount}/home/pi/.ssh/authorized_keys"
 chown 1000:1000 "${sdcard_mount}/home/pi/.ssh/authorized_keys"
 chmod 0600 "${sdcard_mount}/home/pi/.ssh/authorized_keys"
 
+# Clone github repository
+#src_dir="${sdcard_mount}/home/pi/src"
+#mkdir ${src_dir}
+#
+#if [[ ! -e ${src_dir} ]]
+#then
+#    echo "Can't create directory \"${src_dir}\""
+#    exit
+#fi
+# Clone source directory to target
+#git clone "${github_repo}" "${src_dir}/BeholderPi"
+
+# Install discovery service
+echo "Installing Discovery service"
+sender_py="${scripts_path}/services/beholder_discovery_sender.py"
+if [[ ! -e ${sender_py} ]]
+then
+    echo "Can't find the sender script \"${sender_py}\""
+    exit
+fi
+cp -v ${sender_py} "${sdcard_mount}/home/pi/"
+
+sender_service="${scripts_path}/services/beholder_discovery_sender.service"
+if [[ ! -e ${sender_service} ]]
+then
+    echo "Can't find the sender service file \"${sender_service}\""
+    exit
+fi
+cp -v ${sender_service} "${sdcard_mount}/etc/systemd/user/"
+
+# Unit files are enabled by symlinking the unit file to a target.wants directory
+echo "Creating symlink to enable discovery service unit file"
+ln -s -v "/etc/systemd/user/beholder_discovery_sender.service" "${sdcard_mount}/etc/systemd/system/multi-user.target.wants/beholder_discovery_sender.service"
+
+# Done modifying image
+echo ""
+echo "Preparation done, copying image"
 umount "${sdcard_mount}"
-new_name="${extracted_image%.*}-ssh-enabled.img"
+new_name="${extracted_image%.*}-ansible-ready.img"
 cp -v "${extracted_image}" "${new_name}"
 
+echo ""
 lsblk
 
 echo ""
 echo "Now you can burn the disk using something like:"
-echo "      dd bs=4M status=progress if=${new_name} of=/dev/mmcblk????"
+echo "      dd bs=4M status=progress if=${new_name} of=/dev/sdX??"
 echo ""
